BUG: camList doesn't work how I want it to
Verified build builds after some quick fixes.  Found that my hasCameras() method doesn't work right now so will need to adjust it in the future (disabled current usage of it).  Also confirmed that some of the behavior works with the program as the fitness value of best individual goes *up* when running across generations.  Given that the laser isn't running this at least confirms that the new individuals aren't removing or corrupting the best and that it appears to be working when dealing with random individuals.