Fixes to operable version again
Resolved a lot of issues that came with trying to cleanup includes again.  End result is improved I think with reduced unneeded includes and stuff where identified and able so far.  Also worked on ImageController to work (and appears to be so!).  Finally identified a cause of the memory leak being forgot deallocation of sorted_temp in the GAs.  As note currently while optimizing the program uses ~160MB of memory.